Add files via upload
This is the report I submitted as an assignment-project. I have used the shared codes in the same repository for the analysis I did in this report. I would like to mention that the analysis does not follow any standard analysis procedures. Due to lack of time, I did not manage to add other details and/or to do the analysis much better way.